Properly handle late HTTPExceptions
The code as suggested worked fine only if HTTPExceptions occurred
before the 201 response had been sent from an Action.
The new code uses a lock to robustly determine
whether the exception should be propagated in the
main thread, or in the Action.

This means that late-raised HTTPExceptions should
now show up in the logs properly.