Handle nullable fields properly
The thing description is invalid if we have fields with `allow_none`
enabled.  That's because JSONSchema permits `type` to be a list, which
is the sensible way to represent a field that may be `null` or a value.

I've modified our marshmallow_jsonschema library to
expand this into one duplicated sub-schema for each type in the
list, combined using `oneOf`.  This is allowed by the thing description
(and, as far as I know, is valid JSON Schema as well).